ml_project/requirements.txt
# local package
-e .

# external requirements
pandas==1.2.4
sklearn>=0.0
scikit-learn==0.22.2
scipy==1.6.3
coverage==5.5
marshmallow-dataclass~=8.4.1
pytest==6.2.1
PyYAML==5.4.1
